Update CardanoSignTransaction method
